backend/requirements.txt
# Core
fastapi>=0.100.0  # This version supports Pydantic v2
uvicorn>=0.22.0,<0.30.0
pydantic>=2.7.0,<3.0.0  # Updated to be compatible with pydantic-settings
pydantic-settings>=2.1,<3.0.0

# LangChain (모듈 분리 기준)
langchain-core>=0.1.0,<0.2.0
langchain-community>=0.0.10,<0.1.0
langchain-huggingface>=0.0.2,<0.1.0
langchain-text-splitters>=0.0.1,<0.1.0

# Database
chromadb>=0.4.18,<0.5.0
pymysql>=1.1.0,<2.0.0
sqlalchemy>=2.0.15,<3.0.0

# Web & API
openai>=1.10.0,<2.0.0
httpx>=0.25.1,<0.26.0
python-dotenv>=1.0.0,<2.0.0

# Document Processing
selenium==4.15.2
beautifulsoup4==4.12.2
trafilatura==1.6.1
pdfminer.six==20231228
spacy==3.7.2
nltk==3.8.1

# Dev
pytest==7.4.3
jinja2==3.1.2
python-multipart==0.0.6
tenacity==8.2.3
numpy==1.26.1

# (Optional)
tqdm
rich
titoken
google-generativeai